+++ ChangeLog.md
@@ -7,6 +7,7 @@
 * Removes extraneous output when unmounting ISOs (Windows 10 1903 only?)
 * Adds Export-LabImage cmdlet to export VHD(X) master/parent images
 * Permits overriding media Id with -CustomId parameter when importing legacy media definitions or media from an external file/Uri
+* Adds 'Latest' property to module info to force latest version download on every configuration run (#367)
 
 ### v0.18.0 ###